+++ src/main/java/com/green/gittest/todolist/ToDoListService.java
@@ -1,9 +1,14 @@
 package com.green.gittest.todolist;
 
+import com.green.gittest.common.CheckMapper;
+import com.green.gittest.common.model.ResultDto;
+import com.green.gittest.common.myexception.ListNotFoundException;
 import com.green.gittest.common.myexception.SignUpException;
+import com.green.gittest.common.myexception.UserNotFoundException;
 import com.green.gittest.todolist.model.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -12,25 +17,47 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ToDoListService {
     private final ToDoListMapper mapper;
+    private final CheckMapper checkMapper;
 
-    public int postToDoList(PostToDoListReq p) {
-        return mapper.postToDoList(p);
+    public ResultDto<PostToDoListRes> postToDoList(PostToDoListReq p) {
+        if(p.getContent() == null || p.getContent().isEmpty()) throw new NullPointerException();
+        mapper.postToDoList(p);
+        PostToDoListRes result = new PostToDoListRes();
+        result.setListId(p.getListId());
+        String msg = "투두리스트 등록이 정상적으로 완료되었습니다.";
+        return ResultDto.resultDto(HttpStatus.OK, "SU", msg, result);
     }
 
-    public List<GetToDoListRes> getToDoList(Long userId) {
-        return mapper.getToDoListByUserIdForRead(userId);
+
+    public ResultDto<List<GetToDoListRes>> getToDoList(Long userId) {
+        if(checkMapper.getUserId(userId) == 0 ) throw new UserNotFoundException();
+        List<GetToDoListRes> list = mapper.getToDoListByUserIdForRead(userId);
+        String msg = "투두리스트 일정을 정상적으로 불러왔습니다.";
+        return ResultDto.resultDto(HttpStatus.OK, "SU", msg, list);
     }
 
-    public int updateToDoList(UpdateToDoListReq p) {
-        return mapper.updateToDoListContent(p);
+    public ResultDto<Integer> updateToDoList(UpdateToDoListReq p) {
+        if(p.getContent() == null || p.getContent().isEmpty()) throw new NullPointerException();
+        if(checkMapper.getUserId(p.getUserId()) == 0 ) throw new UserNotFoundException();
+        int result = mapper.updateToDoListContent(p);
+        String msg = "투두리스트 내용 수정이 정상적으로 완료되었습니다.";
+        if(result == 0) throw new ListNotFoundException();
+        return ResultDto.resultDto(HttpStatus.OK, "SU", msg);
     }
 
-    public int deleteToDoList(Long listId) {
-        return mapper.deleteToDoList(listId);
+    public ResultDto<Integer> deleteToDoList(Long listId) {
+        int result = mapper.deleteToDoList(listId);
+        String msg = "투두리스트 삭제가 정상적으로 완료되었습니다.";
+        if(result == 0) throw new ListNotFoundException();
+        return ResultDto.resultDto(HttpStatus.OK, "SU", msg);
     }
 
-    public int deleteAllTodoList(Long userId){
-        return mapper.deleteAllTodoList(userId);
+    public ResultDto<Integer> deleteAllTodoList(Long userId){
+        if(userId == null) throw new NullPointerException();
+        int result = mapper.deleteAllTodoList(userId);
+        String msg = "완료한 일정 전체 삭제가 완료되었습니다.";
+        if(result == 0) throw new ListNotFoundException("존재하지 않는 유저입니다.");
+        return ResultDto.resultDto(HttpStatus.OK, "SU", msg);
     }
 
 }
